+++ write-rustdoc-hide-lines/src/lib.rs
@@ -1,4 +1,3 @@
 pub mod code_block_definition;
 pub mod formatter;
 pub mod hidden_ranges;
-pub mod utils;
